leaudio: Fix handling removing Available Context Types
When Available Context types are cleared by remote device, Android
should make device INACTIVE and when context types are back, it should
bring it back to ACTIVE. This is used by devics which does not have MCP
or CCP

Since group->GetAvailableContexts() is also taking Streaming contexts
into account, Java layer has no chance to act correctly as described
above if Available Contexts are cleared during streaming. This is
because currently streaming context is still reported as available, as
it was accepted by remote when creating the stream.

Later when remote device autonomously goes into RELEASING state,
Android treats it as an error.

With this patch, Android Bluetooth stack, will double check on the
autonomous release state if there are any available contexts, and if
NOT, then it just notifies Java about missing context types, so Java can
apply special treatment for this use case.

Bug: 314091632
Test: atest bluetooth_le_audio_client_test
Flag: EXEMPT, regression tested with unit tests

Change-Id: Ie2dfa6a37b5f726ca31c05902179eae920080a56