Create directories when dumping snooz
If snoop log are disabled, snooz are used.
Snooz are only written to disk when taking a bugreport.
When starting the stack for the first time (or after a factory reset)
the log folder could not be present and we should first create it

Test: m .
Flag: com.android.bluetooth.flags.snoop_logger_recreate_logs_directory
Bug: 417137007
Bug: 383876267
Change-Id: If3a14d8e1a2bee4b1f6dc9a20c4dec0d98731e04