Clear pointer to freed alarm
When an alarm is freed, it is no longer valid to access that memory. Setting it to null helps to track accidental uses of freed alarms. It is also a good practice to set the freed pointers to null.
`alarm_free()` does the `osi_free()` at the end, so its safe to set the `alarm_t*` to NULL in the caller. Ideal way is to pass the reference of alarm_t*, and set it to NULL in `alarm_free()` itself.

Bug: 414525236
Bug: 415965031
Test: mmm packages/modules/Bluetooth
Flag: com.android.bluetooth.flags.set_ptr_null_after_free
Change-Id: Ia4e70ab0b6b9d1e6107e9d9efc81703366f63b0b