Log apps' transport preferences
Problem: Client and server applications have transport preferences
assocaited with them that can eventually be used in logic to help make
sure we're sending things like notifications down the bearer the server
prefers, etc. This transport is supposed to have some impact on how GATT
operations work in the stack, but we have no visibility into what it is
inside the stack itself.

Solution: Start by piping this to GattService and logging it, so its in
the stack can be be used at some point.

Flag: EXEMPT, logging only
Bug: 382778345
Test: atest com.android.bleutooth.gatt
Change-Id: I22b9cf6d2edce16677b89301c61259a011627851