Bass: Use resume logic for bud reconnection
In the current code, when a bud, which was synced to the broadcast,
reconnects to the phone and other devices from its group are still
synced to the same broadcast, add source is called no matter the
current state. This lead to start broadcast on this bud even if on
another bud is playing unicast because of handover.

This CL fixes that by adding a sink device to the paused sinks and
calling resume if needed (broadcast not suspended nor monitored).

Removed check of already added source as it is handled by resume,
and an update of source is needed in case of no PA/BIS.

Additionally, a retry mechanism will be used here because of the
use of the whole resume logic.

Bug: 417663083
Bug: 417659324
Flag: EXEMPT, day-to-day fix covered by unit tests
Test: atest BassClientServiceTest
Change-Id: I3a7bd0b030d310e52d32133757bc190b9c1a3ddf