Don't use transaction ID directly with apps
Problem: A client request is made up of a connection ID (bearer),
transaction ID (server assigned from native, so client can associate an
eventual accepted request with its response) and a handle (represents
the data being asked about). The domain used for the transaction ID is
local to the bearer in question. In GattService.java, we map the
transaction ID to a <connection ID, handle> pair, using it as a key,
assuming its unique across all devices and bearers when it is not. This
means that, in rare cases, transaction IDs can be overwritten, causing
data to go to the wrong bearer, or even the wrong device!

Solution: Instead of using transaction ID directly to identify a
request, add an extra layer of indirection in the form of an auto
incrementing integer which we'll refer to as a request ID in the handle
map. We'll now map a request ID to that requests _entire_ context, being
<connection Id, transaction Id, handle>, and send _that_ to the server
app to return with their response status and data. We can then use that
to look up the context and send our response to the correct bearer with
the correct transaction ID. Since each request the comes in gets a new
request ID, there's no squashing or stepping on values.

Flag: com.android.bluetooth.flags.gatt_multi_bearer_transactions
Bug: 418050422
Bug: 415380437
Test: atest com.android.bluetooth.gatt
Change-Id: Ifc033982e9f1a3c0c24741c4e8285572ca34c26e