HFP: listen for telephony on background
Telephony manager API are blocking, and in some situation, they can take
a couple secondes (up to 5~7!). It is not conceivable to block the
Bluetooth main thread for that long, as this would have cascade effect.
Instead, putting update on the same BackgroundExecutor thread, that
would do is work independently of Bluetooth turning off or not

Note, the code will still synchronously wait on the startPhoneListener,
to guarantee this is a no-op change when it actually matter to listen.
But it will not wait on stopping the phone listener, as there is no more
usage of the class at shutdown after stopping a final time

Bug: 417747853
Bug: 411223749
Test: m .
Flag: EXEMPT no-op refactor validated by stress testing
Change-Id: I6a945210b4c1c712a346dd59b0cb36d4c4d0243c