leaudio: Fix missing Cancel Audio HAL request while stream release
Intention of this code is to Cancel Audio HAL request when stream is
releasing.
Checking group state is partially correct, but this state changes to
RELEASING when all ASEs are in RELEASING state.

Instead of checking group state, stack should use IsReleasing()
function, which is checking if RELEASING is in progress

Bug: 416397249
Test: atest bluetooth_le_audio_client_test
Flag: EXEMPT, trivial day-to-day fix.

Change-Id: I8b5a74e33e97353e522c93ee7b772818b2239dbb