Reorganize GattService unit tests
Problem: Our GattService implementation contains the implementation for
_both_ the client and server, but the test coverage is lacking, and its
very hard to tell which functions belong to which side when looking at
the existing coverage.

Solution: While it would be nice to separate our client and server
implementations and put well defined boundaries between the two, and do
the same for the tests. We're not ready for that. Instead, lets
reorganize the file to be clear about which tests belong to which side.

Flag: EXEMPT, test only
Bug: 382778345
Test: atest com.android.bluetooth.gatt
Change-Id: I18cd6b003c0a5926e046f7afe0e4c74a876e0993