Remove custom mock on SystemProperties
We are now linking with extended mockito and can mock static class
directly. We no longer need to have a dedicated utiliy class to do mock
injection.
Also cleanup some usage with setter that are not needed if we can
instead set the device in the correct condition for testing first

Bug: 311772251
Test: atest BluetoothJavaUnitTEst
Flag: TEST_ONLY
Change-Id: I317010149f885775c9159ac049ce5620a91bc438