avatar: Resolve Asha flakiness issues
FactoryReset was the source of some amount of flakiness
for AshaTest as PandoraClient was not waiting synchronously
for the channel to be shutdown after FactoryReset, resulting
in the channel being reused for RPCs during test execution,
later causing SHUTDOWN failures.

Remove some of the ASHA tests that are in fact only testing
basic LE capabilities and no ASHA features.

Bug: 379847877
Test: atest avatar.AshaTest
Flag: EXEMPT, test change
Change-Id: I3de998960fdf29011a86fdee836067f76b36824a