le_audio/client.cc: do not set group without conv ctx active for call
If device does not have CONVERSATIONAL in available contexts, it can't
be configured. Previously, group was set to active which lead to
reconfiguration and constant loop of attempt to configure the device,
followed by disconnection.

With this patch, group will never attempt these steps and never will be
set active withouth CONVERSATIONAL context, for call.

Bug: 407641311
Flag: EXEMPT, day-to-day bugfix covered with unit test
Test: atest UnicastTest#GroupSetActive_DuringPhoneCall_NoConversational
Change-Id: I1acef9319543874b11d28fd3a254bfbbcd3c483f