ActiveDeviceManager: only set lead device as active LEA device
When a LEA device is conencted, there is an attempt to set
it as active, always. This is not a valid behavior for two reasons:
1. it fails when before second device connects, the other group connects.
  This makes first group active again, instead of new one.
2. We only allow to call LeAudioService#setActiveDevice on lead device.
  This means we are wasting time in handleLeAudioConnected, just to
  call setLeAudioActiveDevice and exit there early.

This patch checks if connected device is lead device, and sets it
as active when it is, resolving both issues.

Adjusted some unittests so they use appropriate devices for
their scenario

Bug: 364441697
Flag: EXEMPT, day to day bugfix covered with unit test
Test: atest ActiveDeviceManagerTest#leAudioSecondDeviceConnectedAfterOtherGroupConnected_setActive
Change-Id: Id40ec748bbe85395396733694a59a9a14c8a6908