btm/iso:  Fix possible race when setting data path.
This is fix handling RemoveIsoDataPath when is called before
SetupIsoPath is not yet completed.

It can happen when Disconnect Complete Event occure just in the middle
of the SetupIsoDataPath operations.
Stack is handling then Disconnect Complete Event in the mainloop and
tries to RemoteIsoDataPath which leads to assert.

Bug: 416192597
Flag: EXEMPT, day-to-day fix, regression handled with unit test
Test: atest net_test_btm_iso
Change-Id: I7970a14b86e6a3e4a6a1fe804367960adac608a4