Save GATT DB cache for bonded devices
Currently, the GATT DB cache is saved only when
discovery is finished AND the device is bonded.

This means that if bonding occurs after discovery has completed,
the GATT DB cache is not saved.
This change addresses that by ensuring the DB cache is saved
upon successful bonding, regardless of the discovery state.

Flag: com.android.bluetooth.flags.save_cache_for_bonded_device
Bug: 417859848
Bug: 414318763
Test: atest bluetooth_test_gd_unit
Change-Id: I62c6d21b9cb743687b3215a9dbde8af2626af7f6