Delete ProfileServiceTest entirely
This test, realistically, tests nothing.
It only forces scenarios within itself to pass
and forces other classes to follow. So not only is
it not a "unit test", testing a unit, but it forces real
classes to shape themselves to be fitting here.

Classes that work in good, clean code were disabled
in `manuallyDisabled` or `excludedProfiles` due to
not adhering to the above mentioned _fake_ flows.

This opens up the possibility of ProfileService dependencies such as
their NativeInterface to be proper final fields as this test
was a unintentionally for them.

Bug: 418183047
Test: atest AdapterServiceTest
Flag: EXEMPT test only
Change-Id: Ie4493818fc321fb65757f962ef0651331cbac270