Remove singleton from BluetoothKeystoreNativeInterface
Set the field `BluetoothKeystoreService mBluetoothKeystoreService` as final.

This improves testability and simplifies dependency management.

THis also leads to a tradeoff. It adds the benefit of a final
`BluetoothKeystoreService` field in `AdapterService` while losing the
final-ness of `mIsCommonCriteriaMode` within the class itself.
I believe it is safer to have a final object field vs a boolean and hence: A net win.

Bug: 418183047
Test: atest AdapterServiceTest
Test: atest BluetoothKeystoreServiceTest
Flag: EXEMPT refactor no-op
Change-Id: Iae06e276665f5d2d53c1ec7e50c41921cb35f8af