Create new ScanRadioStats (Decoupling from AppScanStats)
Almost all static logic from AppScanStats regarding ScanRadio is
extracted to ScanRadioStats.

This allows better single responsibility and easier unit testing.

There is a lot of stuff going on that may or may not belong. But,
I have tried my best to move with as little logic change as possible.
Every line change is aimed at to be only required changes and no optional
improvements. Those can be in a follow up.

Bug: 394958283
Bug: 420548930
Bug: 421138363
Test: atest AppScanStatsTest
Test: atest ScanControllerTest
Test: atest ScanManagerTest
Flag: EXEMPT refactor no-op
Change-Id: I1c48f4f2979ce7cda2e85a16acec5f60201b5cda