Migrate all* OPP instrumentation tests to unit tests
* Close to all.

Adds an additional 24 tests to atest com.android.bluetooth.opp
- From 106 -> 130

Some tests rewritten with OPP util methods.
Only 2 tests had to be deleted.

- BluetoothOppService_CleanupTest did nothing that BluetoothOppServiceTest
  doesn't.
Also helps get rid of the constructor in AdapterService that served no purpose
within it which is a win.

- IncomingFileConfirmActivityTest.java#
  onTimeout_broadcastUserConfirmationTimeoutAction_sendDismissTimeoutDialogMessage
Nice to have but a basic unit test that was doing sendBroadcast of
USER_CONFIRMATION_TIMEOUT_ACTION and verifying nothing happened onTimeout.
The logic was trivial and a tradeoff for having the rest of the class available.

Bug: 419891918
Test: atest com.android.bluetooth.opp
Flag: EXEMPT test only
Change-Id: Ia3cc7271a1481ed32bb7603e9b7a7975a8429d80