[SAMSUNG] Fix service uuids not cleared.
If the remote device disconnects without responding to the SDP request,
the DUT retries the SDP. However, the connection is already disconnected,
so avrc_cb.service_uuid remains in avrc_sdp.cc.
After that, AVRC_FindService will fail with AVRC_NO_RESOURCES until the next reboot.

If a connection is disconnected on the incoming connection,
SDP request should not be made, and the service_uuid must be initialized when BT is turned off.

Bug: 420850916
Flag: EXEMPT trivial
Test: m .
Change-Id: I7c31cbbf218e934a496bb5bba8c44747d9bc10eb