Finalize almost everything in `AdvertiseManager`
Also make `AdvertisingSetDeathRecipient` private. Why was it public 🤔.

This should really be very optimal to do. With flag `advertise_thread`,
enabled, almost all the code in AdvertiseManager is moved to a single thread.

A few crashes have been reported last few dogfoods in `stopAdvertisingSet`
especially. I've taken the liberty and started off with that and then
did the rest of the whole class which should be at worst, a nano gain,
at best, a minimal gain of thread safety and optimization.

Bug: 420596891
Bug: 391508617
Bug: 421138363
Test: atest AdvertiseManagerTest
Test: atest GattServiceTest
Flag: EXEMPT refactor no-op
Change-Id: I3de9b24fe87317399dd818642ec0c2f6db45cdbb