To handle Voice Recognition security exception on Bluetooth process
Description: If the Voice Recognition target application launched by a
VOICE_COMMAND intent applies the activity export option to 'false', and
Voice Recognition is then triggered by a headset.

Basically, it need to be fixed from Application side to changing the
activity export option to true. However, we have added a catch condition
to avoid exceptions in the Bluetooth process.

Bug: 421353374
Test: m comm.google.android.bt
Flag: EXEMPT no-op
Change-Id: I0359a930f5bd638a2ac1b8f6e760b8348f6b1ff4