Replace `getName()` with `@Override toString()` for profile logging
Replaces the custom `getName()` method in `ProfileService` with a
standard override of `toString()` override to provide a string
representation for logging.

Since `AdapterService` moved to checking equality of `GattService` by
id and not by name, this is no longer required both in app code and
in tests.

Bug: 394958283
Test: atest BluetoothJavaUnitTests
Flag: EXEMPT refactor no-op
Change-Id: I2113ee56e581af5e96cf783fc1ff2fcbc9dce038