Move BLUETOOTH_PRIVILEGED permission check to GattServiceBinder for `mRestrictedHandles`
The last permission check. Technically the field `mRestrictedHandles` and
the function `void permissionCheck` belong in GattServiceBinder but
moving them would be exposing a bunch of private class fields from GattService.
I thought it's just better to keep them in the service and delegate the
calls to the Binder to achieve the same purpose of this CL & attached bug.

Bug: 409629399
Test: atest GattServiceTest
Test: atest GattServiceBinderTest
Flag: EXEMPT refactor no-op
Change-Id: Id207d9dae1b7e7f786b61d4c7e435fbde1b69a6c