Speed up PhonebookPullRequestTest
Test runtime on a cuttlefish goes from ~8.4s -> ~4.7s (44% faster).
Calculated by averaging the speeds of 10 runs via
`atest -itd BluetoothJavaUnitTests:PhonebookPullRequestTest --iterations 10`

`MAX_OPS` in `PhonebookPullRequest` is defined as `250` yet the test
was adding 200 + 200 + 250, way exceeding the limit; with no additional benefit.
The test is updated to add MAX_OPS/2 + MAX_OPS aka 250/2 + 250 which
still exceeds the limit (for same purpose whatever it was yet faster).

The test does not even check the limit exceeding or not but I'm keeping it
for now to not change any "logic"...

Bug: 404652736
Test: atest PhonebookPullRequestTest
Flag: EXEMPT test only
Change-Id: I7a296a5722547d3171cf48feaf09a552c72e151a