leaudio: Fix Audio HAL response when dynamic directions are used
When restarting direction from QoS Configured to Streaming was done,
the stack could not respond to the Audio HAL correctly to Resume request,
which could lead to audioserver timeout and crash.
This is because audio_sender_state_ and/or audio_receiver_state_ was no
correctly updated.

Bug: 421013874
Bug: 416397249
Bug: 412540131
Flag: com.android.bluetooth.flags.leaudio_dynamic_direction_opening
Test: atest bluetooth_le_audio_client_test
Change-Id: Ice6e24257de0707df7df6151973ec9992185f5d6