Create Optional getters for all BluetoothProfiles in AdapterService
The getters are defined in the same order found in PROFILE_CONSTRUCTORS.
Which is the alphabetical order of `BluetoothProfile.ID`.

This is done in 1 CL so it prevents merge conflicts as part of multiple pending
CLs that were undesirably being linked together.

Bug: 422543753
Test: atest AdapterServiceTest
Flag: com.android.bluetooth.flags.adapter_service_profiles_use_optional
Change-Id: I6c9b1fc340fb192b2418d228e7da27aa736beb33