Bass: Add getting metadata for autonomously synced broadcasts
If the remote autonomously syncs to broadcast or the remote with
an already synced broadcast connects to DUT, then, without this CL,
Broadcast Assistant cannot control the synced source on the remote.
This causes, i.e., an issue with starting unicast. The remote has no
MEDIA context, so the broadcast needs to be stopped, which is not
possible without metadata.

This CL fixes that by syncing to broadcast and update metadata in such
cases. The scanner is enabled in the background and did not influence
any other operations.

Timeout for syncing in this scenario is set to 1 minute.

Notifications about start/stop scanning or sync found/lost are sent
only on foreground scanning.

Bug: 422398083
Bug: 422399972
Flag: com.android.bluetooth.flags.leaudio_broadcast_fix_autonomous_source_adding
Test: atest BassClientServiceTest BassClientStateMachineTest
Test: atest BluetoothLeBroadcastAssistantTest
Test: manual tests with autonomous source adding
Change-Id: I43005887d2e2a86d5ad41c4d623680709fad3642