Follow-up on getEncryptionStatus
We should not have a file to define the inner parcel of the encryption
status. We should not redefined the structure in 3 locations. We should
not have different names for the same thing.

Also, run the formatter on the rest of the repo

Bug: 419143231
Test: m .
Flag: com.android.bluetooth.flags.link_status_api
Change-Id: I7c62d3e1a5ef1880259c0b768aebf51193944b4a