Remove singleton from `SdpManagerNativeInterface`
This improves testability and makes the ownership of SdpManagerNativeInterface explicit.

Set the field `SdpManager mSdpManager` in it as final.

Also remove the `mNativeAvailable` boolean and the corresponding `isAvailable()` check
to simplify the object's existence, managed by AdapterService and exposed via an Optional.

Moving to non-singleton, Optional added safety checks to
*3* locations which were previously not checking the availability.
Now, the natural empty check of Optional will provide the benefit
of not calling native interface functions in the unlikely chance of
when they are not available.

Bug: 418183047
Test: atest AdapterServiceTest
Test: atest BluetoothPbapServiceTest
Test: atest DipTest
Flag: EXEMPT refactor no-op
Change-Id: Ia2421dc2ddf3b8de1429626e6056f212009fda70