VC: Check device readiness and need for update before forwarding op
If VC stack forwards operation to a device that has such value already
set, then no notification change arrives. That will cause to stuck with
a pending operation without the possibility of recovery. This CL fixes
that by checking if such an operation is required.

Bug: 422337238
Flag: EXEMPT, day-to-day fix covered by unit tests
Test: atest --host bluetooth_vc_test
Change-Id: I03c315163972e857dd369bc3fc45b4ffbfc90bcb