Inject `NotificationManager` into `BluetoothPbapService`
Possible now due to the all-new `constructProfile`! And a great way
to avoid `, null` in primary constructors.

This also allows lower the dependency of `BluetoothPbapService` on being
a `ContextWrapper`.

Note that the passing down of `mNotificationManager` to `PbapStateMachine`
is not new nor tied to the flag and just a convenience change.

Bug: 422543753
Bug: 417622882
Test: atest BluetoothPbapServiceTest
Test: atest PbapStateMachineTest
Flag: com.android.bluetooth.flags.adapter_service_profiles_use_optional
Change-Id: I395a9dfbc07252b27ac8a92e6c9d88b53e55ab3e