Remove redundant null checks in HeadsetService for 2 final fields
The field `mAdapterService` is final and guaranteed to be non-null

The field `mSystemInterface` is final in `HeadsetService`.
Inside `HeadsetSystemInterface`, the field `mAudioManager` is also final.
Crucially, the HeadsetService constructor calls
`mSystemInterface.getAudioManager().registerAudioDeviceCallback(...)`.
If `getAudioManager()`` returned null, the service would have crashed
with a `NullPointerException` during initialization.
Therefore, if the service is running and the callback can be invoked,
`getAudioManager()` must have returned a non-null object.

Bug: 394958283
Test: atest HeadsetServiceTest
Flag: EXEMPT refactor no-op
Change-Id: Icd752a046c82622a8cb2746757b0b18f2a9c1500