Dead code removal - BLE_CONNECTABLE state
Looks like this is some remains of code that would start advertising when device is in connectable mode.

It is for certain non-functional because it claims to send the legacy HCI advertising commands, and this is forbidden by the spec, when the rest of stack uses extended advertising.

It was incorrectly used in btif_dm.cc. ble_evt_type doesn't have anything to do with wether the advertisement is connectable.

The btm_ble_set_discoverability is next for removal, but it's not so easy to prove it's not used.

Test: mma -j68
Flag: EXEMPT, no logical change
Bug: 415832385
Change-Id: I5fa9080c0a8e6cbb8713ddf7bf2f4d083558d348