leaudio: Extend suspend timeout during the call
Whenever Audio HAL Calls suspend on the Encoding and Decoding session,
Bluetooth stack keeps CISes for a 500ms in case, Audio HAL decides to
Resume the stream withing this time.
There used to be 5 sec timeout before but that turned out to be too
long and could cause IOP issues with remote devices having no idea why
we keep CISes that log e.g. after stream is ended.

The main use case is when user is generating touch sounds -
basically Bluetooth stack does not want to stops and start stream again
as this is costly.

The other use of this timeout is when Audio HAL for some its internal
reasons, needs to call Suspend/Resume in a sequences. This  is is
happening e.g. when there is device switch or other. Without this
timeout, we would spent lot of time on dropping and re-creating stream
on each Suspend/Resume.

For this second use case, when the switch happens during the phone call,
it turns out that 500ms is too short and Bluetooth is losing lot of time
on reduntant reconfiguration while Audio HAL is informing by Audio
Framework about device switch.

This CL extends Suspend Timeout to 2 sec when Suspend is happening
during the call. This helps to decrease time for switch when Audio HAL
is doing multiple Suspend/Resumes calls.

Bug: 385262729
Bug: 438399588
Flag: com.android.bluetooth.flags.leaudio_improve_switching_le_audio_devices
Test: atest bluetooth_le_audio_client_test
Test: Manual testing

Change-Id: I6a08f0e8a816754e8855cc5b17e4f7dbc18b6376