Improve AdapterBinder#killBluetoothProcess logging and final modifier
While looking into b/371056313, I could not see if this still happens
when the new flag `Flags.killInsteadOfExit()` is used. So I believe
logging the killType in the `Log.wtf` can help understand if this
issue still happens with it enabled or is it resolved. At least until
the flag is fully rolled out.

Bug: 339553092
Bug: 371056313
Test: atest CtsBluetoothTestCases
Flag: com.android.bluetooth.flags.kill_instead_of_exit
Change-Id: I35825017ec9b82c39a1d137d068e51fe3276e5c6