Reuse the same `ScanClient` instance for flushPendingBatchResults
Previously, a new `ScanClient` instance was created with only the
`scannerId` considered for flushing for every flush call. This was
leading to wrong logging as the default `ScanClient` could have a
different `ScanSettings` than the requested `ScanClient`'s ScanSettings.

Fortunately, this was not breaking anything as its only ever considered
for its `client.mScannerId`. Hence, the object was created and ignored
for no reason other than its int `mScannerId`.

This change now attempts to retrieve the same `ScanClient` object.

Bug: 394958283
Test: atest ScanControllerTest
Flag: EXEMPT refactor no-op
Change-Id: I7ef70bb37f2f54f364bdcdc2113d734d47ab4ca9