Fix redundant ScanSettings creation
Changing `>` to `>=`, to correctly handle cases where the report delay
is exactly equal to the floor.

In `ScanController#enforceReportDelayFloor`, logs are added to trace how
the report delay floor is evaluated and applied. This makes it clear
whether the original delay is used, or if it's enforced by the
configured floor value.

In `ScanManager#getDeliveryMode`, logs are added to explain why a
specific delivery mode (`IMMEDIATE`, `BATCH`, or `ON_FOUND_LOST`) is
chosen for a scan client based on its settings.

These logs only get printed once so its not spammy and safe to add.

Bug: 394958283
Bug: 428791432
Test: atest com.android.bluetooth.le_scan
Flag: EXEMPT refactor no-op
Change-Id: I81aa0e979904e25663a1d15b4c4129afb68f8841