LeAudio: Introduce PAC service
This CL introduces a Published Audio Capabilities GATT server module,
which is about to be instantiated by a higher layer of LE Audio Unicast
Receiver implementation.

It serves as an abstraction layer on top of the low-level GATT
interface and provides a relatively simple interface.

Bug: 428719427
Bug: 424275371
Test: atest bluetooth_le_audio_server_test
Flag: com.android.bluetooth.flags.leaudio_peripheral_feature
Change-Id: Ic99b74854799f8d28c8cc8fb1471d57244a7f844