Add tracking for batch alarm scheduling per app basis
This is an important figure as it causes a device wakeup and may have
significant battery impact if done too frequently. This change
introduces a new metric to track the number of times a batch scan alarm
is scheduled per application.

The new `mScheduledBatchAlarmCount` metric is also exposed in the
dumpsys output for `AppScanStats`, improving visibility for debugging.

Bug: 427832915
Bug: 428791432
Test: m BluetoothJavaUnitTests
Flag: EXEMPT logging and dumpsys only
Change-Id: I777748efc564fecdcdd7a395cad7ca26f3f7087a