Fix le_impl not unregistering scheduler
in aosp/1952299 mechanism was introduced that would disable callbacks
when the connection object, that is associated with callbacks, is
deleted. Unforunately, as part of this fix, a bug was introduced - if
keeper of this connection object, shim/acl.cc invalidates the callbacks,
it would also prevent the unregistration of round robin scheduler.
If connection is established again on same handle, this would result in
another registation attempt, which triggers an assert and crash.

Test: mma -j32
Bug: 384354768
Bug: 213537005
Flag: EXEMPT, daily bugfix
Change-Id: I389fd32e0f45bb5b55889254cec9b078f3df2107