Fix scan result stats for batched scans
The total number of scan results reported in the Bluetooth dumpsys was
often incorrect or zero, particularly for applications using batch
scanning.

This was because the result counter in `AppScanStats` was only
incremented for non-batched scans. For hardware-batched scans delivered
via `onBatchScanResults` or a `PendingIntent`, the results were sent to
the client app, but the statistics were never updated.

Bug: 429936030
Fix: 429936030
Bug: 428791432
Test: m BluetoothJavaUnitTests
Flag: EXEMPT dumpsys only
Change-Id: I765da1c89f064679ca80ff8e8e59865597f9b412