[HFP SCO Management] Fix dialing out intent
If setActiveDevice is called for the current active device, true is returned and the audio fwk callback is not called.  This adds a check to check if the exposed device is the same as the device that is triggering the dialing out intent and starts the dialing out intent immediately.

Bug: 429537412
Test: m com.android.bt
Flag: android.media.audio.sco_managed_by_audio
Change-Id: Ie864f4b5b7aa14e0bfee6fe1ea97546d3020b247