Add request context for prepared write executes
Problem: Preprared writes can be requested by a client, having server
implementations hold all write requests until the client commits or
executes the write. This execution is meant to commit all the writes,
in the order they were received, on the server. This operations is
exposed to server apps via the 'isPrep' field in the various request
callbacks, combined with the onExecuteWrite() callback when the client
is ready to commit. Apps can respond back with a sendResponse() call
upon completion of the prepared write.

While all operations have transaction IDs, the characteristics written
to usually have attribute handles that are exposed to apps, but the
prepared write commit operation itself does not. Despite this, our
APIs used to send a response expect a handle. We currently keep a
mapping of transactions to outward facing request IDs, connection IDs
used for the transaction, and the handle used in the transaction.

The ultimate problem is that we don't create a request ID and context
for prepared write executions.

In the pre-multi-bearer-transaction world, we would log the missing
context and get usually lucky by using the 0 default-value handle and
the first connection ID found associated with the remote device. The
handle is ignored and the connection ID is correct in the case there's
only one bearer. Furthermore, we're also lucky that the default handle
of 0 is also not one within the valid handle space for apps-- 0x0 is
considered invalid by the ATT spec.

In the new world, we drop requests we don't have a record of though.
This is intended and really what we should be doing, but this means
we can't forget to create request contexts!

Solution: Create a request context associated with the prepare write
execute, to map back to the proper connection ID And transaction ID.
Use an explicit 0x0, via a readable constant, to indicate that we're
purposely ignoring the handle value.

Flag: com.android.bluetooth.flags.gatt_multi_bearer_transactions
Bug: 429238407
Test: atest com.android.bluetooth.gatt.GattServiceTest
Change-Id: I00708a9af7b8ede5bcbcc3b7a09702a795718114