leaudio: Avoid spare stream reconfiguration with some headsets.
Some of the headset configuration plus Android touch sounds might cause
spare Stream reconfiguration which might be not nice for the user e.g.
When pause and resume video on YT.

It might happen especially when remote device does not have SOUNDEFFECTS
in the Supported Context type as this cause Android
sending UNSPECIFIED context when SoundEffect is generated.

Bug: 424015473
Flag: EXEMPT, day-to-day fix, regression tested with unit tests
Test: atest bluetooth_leaudio_client_test

Change-Id: I85ba6e20d77d7e7da4e3ddffa7ce62f942a4442d