Factor out transaction ID increments
Problem: When enqueuing a command, we need to assign it a transaction ID
that's associated with the command. Each transport has its own control
block with a transaction ID integer space that's meant to be incremented
through and wrapped around. The code helping do this is copying the
increment code in both the if/else blocks when it doesn't need to

Solution: Factor out the transaction ID increment logic and simplify the
if/else block

Flag: EXEMPT PURE_REFACTOR
Bug: 415382320
Test: m com.android.bt
Change-Id: I9c6aa134fdad450999423dc3b70ec35c36f0ed55