Use advertising slots that are marked not in_use.
Although we do check if the advertising map contains the advertiser ID
before each access, it might be possible to have a race condition where
the advertising slot is removed after the check and then recreated on
map access. The in_use flag is set to false by default and only set to
true when allocating the advertiser, so using a slot that has in_use as
false is safe. For additional safety we are reinitializing the slot
here.

Also moves count and find checks to contains, except where the return
value of find is subsequently used after the check.

Bug: 405155501
Flag: EXEMPT, trivial
Test: m Bluetooth
Change-Id: I8ea0ab1f7bbada1de0e7c19252d858b22b2f9457