Introduce lower_tester class with a capability to intercept and modify BT packets
This CL introduces a LowerTester class with a capability to intercept and modify Bluetooth packets. The class will be used in the future to test Bluetooth interoperability scenarios, such as:
- Simulate a faulty device: a device that sends malformed packets or does not respond in time
- Simulate a device with different capabilities: a device that supports a subset of Bluetooth features
- Simulate a device with different behavior: a device that behaves differently than the specification
- Test for security vulnerabilities: e.g., by injecting malicious packets
- Test for interoperability issues: by injecting unexpected packets or by modifying expected packets

The LowerTester class can be instantiated with 2 BT addresses, port for socket and a function that will provide instructions for intercepted packet.

The class has the following main functions:
- Receive packet over socket
- Intercept packets
- Send packet to destination device (either device to device pair or back to the sender device)

The class uses the following classes:
- PacketInstructions: to define how the intercepted packets should be modified or handled
- PacketView: to represent the packet received or to be sent
- TesterUtil: to provide some utility methods

This class is not yet used in the testing framework, the next CL will introduce a test case using it.

Bug: 406379764
Flag: EXEMPT Test
Test: mmm packages/modules/Bluetooth
Change-Id: I6055553b671726b91a62cb0fc51a18dc023af96d